Several major changes in grid definitions
The way grids are manipulated has changed. The most relevant change is that getGrid now returns the grid-cell resolution via resX and resY attributes, rather than within the x and y components of the output list. This way it is possible to ascertain by the interpolator whether the grid is inherited or defined explicitly by the user, hence modifying its behaviour accordingly